Tweaks to the max light loadout.
Fixed a bug in figuring out which exotics to use that was pointed out by @thesamprice.
Added a fudge factor to prefer items that are already on your character if they have the same level as something else.
Exclude items that require a higher level than your current level as candidates for the loadout.
Handle the case where a character has no applicable items of a given type (like Artifact).