Avoid reloading stores unless changing equipped item.
This is a second try at 8f3d113463c3d8f79af702080ff9556530c7da6c that
avoids the problems in #353 (by correctly wiring a catch handler). Now,
we will only make the call to reload info if the active equipment is
being changed (equipped or dequipped), so we can see the new light
level. Otherwise, we can simply trust that the move worked as we
intended. This should speed up moves a bit.