Reload stores before trying to move an item to a full inventory.
This gives a chance for the move to succeed if our view of inventory is
out of date (due to trashing or moving an item in game). As part of
this, I had to switch the dimStores directive to listen to the
dim-stores-updated message, which should also make it reload anytime
anyone gets stores from Bungie.