Added basic implementation of viewing possible int/dis/str rolls for gear for any given character
Initial implementation using some of the commented code in MinMax.

This is done using only the best gear pieces in terms of int/dis/str.

The next version should include a mechanism to lock specific armor
pieces in, for example, a user might want to use a specific exotic and
get the best int/dis/str values.