Added sorting and drag-and-drop locking functionality
Users can now drag-and-drop items to lock items and therefore only
combinations with the locked items included will be generated.

Locked items can be removed by clicking the red x on the top left of the
items.

The armor sets will now show which bonus perk to use in order to get the
values shown on the left.

Ability to view sets that meet certain tiers added with a drop-down
interface.

The sets displayed can be sorted in order of int, dis, and str.

The implementation is still slow. The next update will need to include
some optimizations for this to be usable in the general case.