Update dimApp.i18n.js
Some improvements to the german translation. Unfortunately, there is no rel short translation for "Loadout" in German, as it is needed in this context. "Bewaffnung" only refers to the weapons in the loadout. A semantically correct translation would be "Ausrüstungskonfiguration", but I think that would look pretty bad in the UI.