Idea for handling API keys
just another idea for handling API keys. just opening this as a POC.

the idea is if it’s beta or release then it swaps in the right key, if
it’s a local build it looks at your local storage for your API key, if
there isn’t one set, it asks you to set one.

in the future we could make it a nicer UI with links and more
information, the `prompt()` was just to get the idea across.