Added in CleanWebpackPlugin and better UglifyJS options
Also set up SQL.js to load async using System.import. Because the SQL library is 'rarely' used (only when the manifest changes), and it's actually really large and doesnt minifiy or compress really well, it makes sense to only load it when required. This gives us a dramatic reduction of the main bundle size, saving 2.3MB from the final bundle, which is really really good