CSV-escape the notes field if necessary
CSV escaping is sort of a nightmare. Escaping is easy:

IF it contains , " \r \n
THEN replace " with "" and emit "..." surrounding the fie;d

Parsing is less easy. Some parsers will handle this correctly. Others
won't.

This patch takes a "can't be worse than before" approach to the problem:
rather than just blindly emitting the note "AAA,BBB" without even
*trying* to escape, which of course breaks the entire CSV, at least
escape it properly so that there's *some* hope of the parsing tools
being able to interpret the CSV correctly (and, specifically, the Perks
column which follows the Notes column).

After this patch, adding a Note to an item with the characters AAA",BBB
should result in the CSV Notes field containing the chars "AAA"",BBB"

Fixes #1370, if it passes testing.