Bargaining with flicker.
We'll hold on tight to ratings for weapon rolls so that we don't need to hit the API for scores we already knew about.

This means that score changes won't currently be reflected until the app gets cycled, but shrug emoji.  We can evacuate our local collection on a submission.