Sort items by "Stat roll percent" now sorts weapons as well
This fixes #1732.

Renamed the sort by “Stat roll percent” to sort by “Rating” which now
includes both the armor (quality rating) and weapon dtr rating.

Sort order would then be:
rating -> prim stat -> abc

The only issue with this is that ratings do not come back instantly,
which means DIM re-orders the items after the request is resolved. This
could be resolved with #1762.