Ornament filter
There might be an easier way to determine if something has an ornament,
but I didn’t see it after some minor digging.

`is:ornament` - returns all items with an ornament node.
`is:ornamentmissing` - returns all items with 1 or more missing (or not
unlocked)
`is:ornamentunlocked` - returns all items with 1 or more ornaments
unlocked on the item.

this fixes #1615