Add missing Search bucket filters
We were maintaining a separate list in the search filters, when we have
the bucket information in dimCategory and D2Categories

Now the search filters are built based off of whatever the contents of
the buckets are.

This mostly adds support for missing D2 “Inventory” filters, such as
Consumables, Modifications, and Shaders.