Applying downward pressure on ratings.
We generally don't get a lot of downvotes on items, but we (ironically?) do get more of those on items popular in the meta.

I made the algo a little more sensitive to downvotes - we weight them more heavily, but the weighting we applies dissapates as the number of total votes grows.