change docker network_mode to host
Corrects the traffic directed to the dockerized dim app. When using
`ports` to map host to container, the traffic is sent to the containers
overlay network interface instead of the container's loopback interface.

Webpack-dev-server listens to localhost, which is bound to the
the loopback interface of the container.

By changing the network_mode to host, we bind the container's network
interfaces to the hosts interfaces, making ports exposed by the container
accessible accodig to the hosts network without needing to map ports.

This bug was introduced in the revert request #3155