acquirable has no 'e'
'acquirable' has been spelled correctly and incorrectly in the repository in various places. In particular, shift-clicking in the organizer on the reacquirable checkmark sets the filter to 'is:reacquireable', but 'acquireable' is not a valid filter.

This also updates the translation mapping to use the correct spelling, but only on the English side. I don't speak other languages but I think this is probably a safe assumption.