Use store that matches preloaded loadout classtype
This should resolve an issue where, when opening the loadout optimizer
from the inventory page, on a character that you haven't played last,
the loadout optimizer wouldn't select the correct character.

This logic sets the store to use on the loadout optimizer to the store
that matches the preloaded loadout's class type. If there isn't a
preloaded loadout, then the code behaves as it did before