feat: add is:enhancementready search and allow is:enhancedperk
The new keyword returns all crafted items that have met the level requirements to enhance any currently-equipped non-enhanced perk on the weapon, plus all enhanced weapons that have met the level requirements to enhance any of the columns on the weapon (that haven't already done so).

This moves craftingData into plugSet so that it can be constructed for every plugSet instead of only for buildDefinedSocket cases.